+++ frontend/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "game_ui",
+  "name": "battle_coding",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
