--- frontend/src/App.css
@@ -1,23 +1,0 @@
-/*메인페이지*/
-#GamePage {
-  height: auto !important;
-  background-color: rgb(246, 220, 184);
-}
-
-/*게임배경*/
-.GamePage_background {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  height: auto !important;
-  width: 100%;
-}
-.GamePage_layout {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  height: auto !important;
-  min-height: 0px !important;
-  padding: 0px 30px;
-}
