--- frontend/src/GameCard/GameCard_Data.js
@@ -1,34 +1,0 @@
-const gameData = [
-  {
-    image:
-      "https://search.pstatic.net/common/?src=http%3A%2F%2Fblogfiles.naver.net%2FMjAyMDExMTBfMTEx%2FMDAxNjA0OTgxNTU0MDMw.wwIzlFbWoHoSxdy52fHN2Z7Hyg91fxKq44E1x2sQFvYg.D1jbkHq-Dd_8aOIPQZUwjNmkCzxrw8gYcpXnK99OuFMg.JPEG.511jay%2F%25B4%25D9%25B8%25AE1.JPG&type=sc960_832",
-    title: "예시 이미지1",
-    description: "안녕하세요 예시 이미지1입니다. 반가워요.",
-  },
-  {
-    image:
-      "https://search.pstatic.net/common/?src=http%3A%2F%2Fcafefiles.naver.net%2FMjAxOTEyMDhfMjU5%2FMDAxNTc1Nzg0NTgwMzkz.3VL0gjnzbKhyNowlO7gwPU-rCwhK4YY7Jf3merQZU50g.GeqMcmluZOAKfxaxSL1QR54V7Wdl-rh0IiQA98z5Q8kg.JPEG%2FexternalFile.jpg&type=sc960_832",
-    title: "부경대 로고",
-    description: "내용 설명",
-  },
-  {
-    image:
-      "https://search.pstatic.net/common/?src=http%3A%2F%2Fblogfiles.naver.net%2FMjAyMzExMzBfMTAw%2FMDAxNzAxMjc4NDg4MTgw.SAcwQCidxC5uVVkmVzjkalkXEb298itrOhejtu242Dwg.1zAbhv7KMoF-F3_QWjTU8GZp8uaH6bXmumXRlWw-uukg.PNG.onlyone8907%2F%25B0%25FA%25C0%25E1.png&type=sc960_832",
-    title: "백경뿌공",
-    description: "백경이와 뿌공이입니다",
-  },
-  {
-    image:
-      "https://search.pstatic.net/common/?src=http%3A%2F%2Fimgnews.naver.net%2Fimage%2F001%2F2021%2F12%2F30%2FAKR20211230048500051_01_i_P4_20211230095631814.jpg&type=sc960_832",
-    title: "백경이 뿌공이 그리고 누구세요?",
-    description: "내용 설명",
-  },
-  {
-    image:
-      "https://search.pstatic.net/common/?src=http%3A%2F%2Fblogfiles.naver.net%2FMjAyMDExMjRfNyAg%2FMDAxNjA2MjA0MzYwNjY0.7iKj4JLHsZ-NdpPPastvxxw90es4uhLNP66PCTEB3tEg._r6CwDRP1hZOe64AFM6c-gBKFQhIAXzpNnGG58lSHq0g.JPEG.dbswls0807%2FIMG_0976.jpg&type=sc960_832",
-    title: "부경대행복기숙사",
-    description: "내용 설명",
-  },
-];
-
-export default gameData;
