--- frontend/src/SearchBar/SearchBar.css
@@ -1,38 +1,0 @@
-/*검색창*/
-.SearchBar {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  margin-top: 20px;
-  margin-bottom: 20px;
-}
-.SearchBar_input {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-.SearchBar_input input {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  height: 50px;
-  font-size: 15px;
-  padding: 0 10px;
-  border: none;
-  border-radius: 5px;
-}
-.Search_button {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background-color: #dddddd;
-  height: 50px;
-  font-size: 15px;
-  padding: 0 20px;
-  border: none;
-  border-radius: 5px;
-}
-.Search_button:hover {
-  background-color: #b1b1b1;
-}
