+++ frontend/src/pages/Game/GameHeader/GameHeader_Guest.js
@@ -1,0 +1,24 @@
+import "./GameHeader.css";
+import Logo from "./Logo.png";
+
+function GameHeader() {
+  return (
+    <header className="GameHeader">
+      <div className="GameHeader_background">
+        <div className="HeaderLayout">
+          <img src={Logo} alt="Logo" className="Logo" />
+          <div className="GameHeader_root">
+            <button type="button" className="HeaderButton_login">
+              회원가입
+            </button>
+            <button type="button" className="HeaderButton_login">
+              로그인
+            </button>
+          </div>
+        </div>
+      </div>
+    </header>
+  );
+}
+
+export default GameHeader;
