+++ frontend/src/pages/Game/GameHeader/GameHeader_User.js
@@ -1,14 +1,18 @@
 import "./GameHeader.css";
+import Logo from "./Logo.png";
 
 function GameHeader() {
   return (
     <header className="GameHeader">
       <div className="GameHeader_background">
         <div className="HeaderLayout">
-          <h3>Home</h3>
+          <img src={Logo} alt="Logo" className="Logo" />
           <div className="GameHeader_root">
             <button type="button" className="HeaderButton_login">
-              로그인
+              마이페이지
+            </button>
+            <button type="button" className="HeaderButton_login">
+              로그아웃
             </button>
           </div>
         </div>
