+++ frontend/src/pages/Game/SearchBar/SearchBar.css
@@ -32,6 +32,7 @@
   padding: 0 20px;
   border: none;
   border-radius: 5px;
+  cursor: pointer;
 }
 .Search_button:hover {
   background-color: #b1b1b1;
