+++ frontend/src/pages/Register/components/UserInputForm.scss
@@ -1,19 +1,10 @@
 @import "../../../style/variables.scss";
 
-.loginSubmitForm {
+.regsiterSubmitForm {
   @include flexColumn;
   align-items: center;
   width: 100%;
   max-width: 25rem;
-
-  input {
-    width: 100%;
-    max-width: 23rem;
-    height: 3rem;
-    margin: 1rem 0 0.5rem 0;
-    border: $border;
-    border-radius: 0.5rem;
-  }
 
   button {
     width: 23rem;
